AP_Math: New Layout to make the header file nicer The implementation is now split in a float and double supporting backend. The float backend is very lightweight and should be used for devices without double precision support. The double backend can be used with Linux based platforms but must be explicitely enabled.
