--- libraries/AP_Math/AP_Math.cpp
@@ -1,2 +1,0 @@
-#include "AP_Math.h"
-#include <float.h>
